// src/app/models/dataUser.ts
export interface IId {
    id: number;
}
export class DataUser{
    id: string;
    id_user: IId;
    id_gender: IId;
    id_country: IId;
    identification_card: number;
    
    names: string;
    last_names: string;
    mobile_phone: string;
    birth_date: string;
    image_user: string;
    local_telephone: string;
    // created_at:	string;
}